side-eye-go: fix bss offset calculations
Before this change, we assumed that the bss field of the firstmoduledata
variable pointed to the start of the bss section. In practice, it points
to the start of the *go* data of the bss section. This address is marked
by the existence of the runtime.bss symbol. The backend now properly
sends the base virtual address for this symbol rather than the bss
section so that the proper conversion is made to locate the
runtime.allgs variable.

GitOrigin-RevId: 79564fba26306c72676956493789b4418a3e9e3f