trace-db: inline the binary hash in the process table
The webappx does not currently benefit from binaries being normalized
into their own table and processes sharing a binary. This patch gets rid
of the binaries table, denormalizing the little information it had into
the process table.

Also, switch from the binary id being a UUID to it being a hex string.
This matches how binaries are identified in the rest of the system.

GitOrigin-RevId: 2d1023a0645daab794e08550fa87e85f896ffebd